refactor: address breaking changes in ui badge
- tag was renamed to badge
- transparent prop was renamed to outlined
